Fix a couple of xpath queries in models
This change fixes a couple of queries that were otherwise not parsing:

- Answer texts on lesson pages: The original query was not matching
the actual XML schema of this content. Also in order to exclude data
we don't want some additional logic was added to filter using
<answerformat> values.

- Question bank match question texts: It turns out the original model
queries were unable to surface a known issue in the question bank, and
the reason was due to match options where answer HTML is modeled as
<questiontext> under <matches>. The xpath query was changed to grab
all <questiontext> elements under a question ID to help address this.